adding eigen libaray to ml bundle

+++ core/vm/interpreter.cpp
@@ -2032,8 +2032,8 @@ void StackInterpreter::ProcessNewArray(StackInstr* instr, size_t* &op_stack, lon
   }
 
   size_t* mem = is_float ? 
-    (size_t*)MemoryManager::AllocateArray((long)(static_cast<size_t>(size) + dim + 2), FLOAT_TYPE, op_stack, *stack_pos) :
-    (size_t*)MemoryManager::AllocateArray((long)(static_cast<size_t>(size) + dim + 2), INT_TYPE, op_stack, *stack_pos);
+    (size_t*)MemoryManager::AllocateArray(static_cast<long>(size) + dim + 2, FLOAT_TYPE, op_stack, *stack_pos) :
+    (size_t*)MemoryManager::AllocateArray(static_cast<long>(size) + dim + 2, INT_TYPE, op_stack, *stack_pos);
   mem[0] = size;
   mem[1] = dim;
